requirements.txt
PyQt6>=6.6.0
Jinja2>=3.1.0
PyYAML>=6.0
Pydantic>=2.5.0
